requirements.txt
Flask==2.2.5
Pillow==10.0.0
Werkzeug==2.2.5
